notifications_api/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic-settings ==2.4.0
aio-pika==9.4.3
pika==1.3.2
httpx==0.27.2
